remove hook package; hooks are not really important at this stage
The new cmd/jujuc/server package contains (so far) the Context type; it
defines the server-side state with which proxied tool invocations can
interact. It is also responsible for running hooks, but that's because it
is the source of truth for much of the hook's desired environment; I'm not
*sure* RunHook will always be attached to Context, but I think that's
something to decide once we have a client.

Context shares some features with the existing hook context classes in
Python, but should not be confused with them. A Context holds all the data
that is needed to execute any jujuc tool [0], plus an Id which is not yet
used directly. It is not concerned with those variables that are purely
related to the actual execution of the hook (ie socketPath and charmDir);
therefore, those are passed separately to RunHook.

[0] Not strictly true, yet; we'll add data when we handle relation contexts,but that's all in the future for now :).

R=rog, niemeyer
CC=
https://codereview.appspot.com/5832045