Try to set up a test for the commandline. Not working yet.
The general outline of the test seems ok. Set up a dummy env as the source,
override the default so that we use the dummy environment as the source.
Create an env that uses the dummy as the target, and run sync-tools, asserting
that the tools get copied to the target.
However, dummy env doesn't seem to have a private bucket. It is public only.
And there is only 1 dummy env, so you can't just create another on to
copy into... :(