Do all of the status requests in parallel.
This is a huge performance improvement. It turns the linear
request of 700 machines into parallel requests for the same, and
similarly parallelizes the requests for unit and service information.

The prefetching of being keys saves us about 50s (down to ~200ms), and
the parallel loading of machines saves us about the same.

Total time is down from 113s down to 6.7s.

Memory consumption is up slightly, 16.428MB up from 13.836MB. This depends
a bit on how much is actually done in parallel, but the tradeoff is
fairly obviously worthwhile.