startup hook for MachineAgent.
When we start the MachineAgent we must have LXC or it crashes in a loop.
This adds 2 main modules a helper for running 'apt-get' in utils/apt
and a cmd/jujud/upgradevalidation.go.
The MachineAgent then always calls EnsureWeHaveLXC on startup.
Because we indirect through containers/lxc all of that code just works
because things that start MachineAgent in the test suite use the lxc.TestSuite
to mock out actually calling lxc.