juju status: add service/unit filters
Any positional arguments after juju status are
now interpreted as filters for either service or
unit. If the filter includes a '/', then it is
considered a unit filter, else a service filter.

If all units of a service are filtered out, then
the service is filtered out; if all units of a
machine are filtered out, the machine is filtered
out (except to fill out the container hierarchy).

Filters are implemented using using shell file
name pattern matching, like in pyjuju. See:
    http://golang.org/pkg/path/#Match