all: implements Environ.SanityCheckConstraints
This introduces a new method on environs.Environ,
SanityCheckConstraints. This method will check
if constraints can possibly be met by the environment.
This will be used to check if an environment is
capable of creating an instance with the specified
constraints.

If the method returns nil, that does not necessarily
mean that the constraints are satisfied; if a non-nil
error is returned, then it means that the constraints
are definitely not satisfiable.

The ec2, azure and openstack Environ implementatons
will now disallow creating containers. When the work
to support addressability is done, this needs to be
changed. Currently, maas and local let everything
through.

cmd/juju is updated so that add-machine, and
deploy/add-unit --to call the new SanityCheckConstraints
method.