Fix bug #1253639, restore compatibility for ssh/scp.
We added Client.PublicAddress in 1.17, so it isn't available when contacting a 1.16 server.
I tested that this functionally matches the old code. We don't do exactly the same
waiting for the InstanceId and then WaitDNSName call that we used to, but it uses
the current code in the API server for instance.PublicAddress.

I put it so it shared the 'wait up to 5s for an IP address' loop.