Finish off bug #1253631.
Pull DestroyMachines and MachineConfig into statecmd so that we can
have them shareable in a location that we know we want to remove eventually.
Trigger the fallback code inside of 'juju add-machine' as appropriate.

This was tested against a 1.16.3 server, and it shows that falling back to
manual provisioning works.