Make bootstrap machine jobs configurable
This addresses a long-standing TODO in
cmd/jujud to making bootstrap machine
jobs configurable. If agent.BootstrapJobs
is present in the agent configuration
key/value map, it will be used instead
of the defaults (the existing hard-coded
values).

The local provider is updated to set
agent.BootstrapJobs to be {JobManageEnviron}.

Fixes lp:1273933