Add ensure-ha command, EnsureAvailability client API
We introduce a new ensure-ha juju subcommand, which
calls the new EnsureAvailability client API. This
is the CLI for ensuring high availability of Juju
state servers.

Note: I have not exposed the ensure-ha command yet,
as there is still work to be done in the agents to
properly support HA state servers.