Fix bug #1311083, have a sane default for -n
'juju ensure-availability' was requiring that you specify -n3, but it doesn't actually
make sense to make it required and not a positional argument. If someone is actually
running the command, the almost definitely want 3, so let's give them a sane default.

The code was explicitly set up to not work this way, but I think it is a better default.