Init EnvCommandBase.EnvName from envs.yaml
This change brings 1.18 in line with trunk,
where EnvCommandBase.EnvName is initialised
to the default environment if current-environment
does not exist and neither "-e" nor $JUJU_ENV
have been specified.

The primary change is in moving environcommand.go
to cmd/envcmd, and adding an Init method that
reads from environments.yaml as a last resort.
All commands that embed EnvCommandBase are updated
to call this method, and various tests are updated
as a result.

Fixes lp:1316869