Invert envcmd relationship
We introduce envcmd.EnvironCommand, and envcmd.Wrap.
EnvironCommand extends cmd.Command with a SetEnvName
method; Wrap takes an EnvironCommand and returns a
Command with the common "-e" flag, and calls SetEnvName
with the environment name prior to the wrapped command's
Init method being invoked. If the environment cannot be
determined, the command will error early.