Fix/remove skipped tests
This change addresses four tests
that were previously skipped due
to being unreliable or untestable
on the bot.

The ec2 test is no longer relevant,
as the test it was masking is no
longer in the source tree.

The keymanager test should be an
integration test (it requires
ssh-import-id, thus network access)
so I removed it.

The cmd/juju tests were calling
"juju bootstrap"; I've changed them
to call "juju env" instead, so there
will be no risk of side-effects or
network access.