Improve the sorting algorithm
The following removes the math.MaxInt64, as we have failure to build on
32-bit architectures. Instead the sorting algorithm now performs one
sort over the indexed collections, first by checking the distance
between two strings and then if they're the same distance, falling back
onto the name.

This should improve the performance as you're only looping over once
instead of twice as previous algorithm did, which is a nice bonus!