Machine context
The following allows us to talk about the types of formatting directives
in terms of their intended output. With these changes we can create
better high level patterns when in the juju code base, ones that
understand if the output is directed at a machine or a human.

We should then be able to reason easily about the intention of the
output type, such that we can change the interaction between the
operator (script or human) depending on what they ask for.

If a machine asks for json|yaml, there should be zero interactive
behaviour. The opposite can then be deduced about human interaction,
where we plausibly want to have more interactivity to ease steep
learning curves.

With the following code changes, we add a new member to the context
struct, one that allows us to know about the type of output that could
be happening. The machine method tells us that a operator/user has
requested a machine format and we should do our very best to honour that
by not outputting stdout/stderr messages that aren't in the correct
format.