Handle stderr vs stdout
When attempting to handle stdout vs stderr for formatting directives it
turns out that it's none-trivial and we should be really consistent with
where our stdout and stderr messages go.

With that in mind, the changes here ensure that when we encounter an
error for a machine format, the EMPTY value for each format is correctly
rendered to the STDOUT and not STDERR.

stderr is useful for error messages and until we have an appropriate way
of putting logs for CLI commands, we should use stderr for constructive
information that the consumer can use.

For scripting, stderr can always be redirected 2>/dev/null or to any
file so they can view what went wrong at a later date.