Log out stderr
So previous PRs attempted to ensure we logged the empty value to stdout
and forsake all errors to stderr. This was a miss step, infact we should
still print to stderr as it is extremely useful.

The way juju currently operates with stderr is very opinionated and we
should be very consistent with those opinions, failure to do that breaks
a lot of commands in juju directly.

The following commit fixes the issue by always dumping out to stderr
even when we have an empty value.

In the future there should be a better place to put that, be we live in
the present and that day isn't here yet. But we should plan for it!