Remove the silent error
The following changes how we handle silent errors returned back from the
handle error function. Originally it was to identify when the output was
written to stdout and therefore didn't need further processing. This is
no longer the case, we do want to handle all errors and pipe them
correctly to stderr all the time.