Update comments
Make it clear in the comments what the intention of the change to the
code does. Handling the error from finding the machine format is bubbled
up to the user and we dump the error stack, so they can attempt to debug
the issue.

Hiding the handler error and not showing the empty value could lead to
confusion, hance why we "try" to do both.