Render all documentation
The following changes the documentation command to use io.Writer
instead of the bufio.Writer. This means that we can then just use
the stdout for writing the documentation output. As the bufio buffer
is actually rather small, only 4096 bytes long. We clearly don't
blow past that during the individual files, but we do when dumping
out the documentation.